Search on key down
Searching on key down prevents capturing key events dispatched when
selecting autocomplete items using tab and return. Since we only capture
key events while the input is focused, the key up event on item
selection is simply not captured and does not need to be handled as a
special case.