Fix dropdown tab navigation
Visibility of the autocomplete dropdown depends on the currently focused
element. If neither the input, nor any autocomplete item is focused, the
dropdown is not displayed.

If the input field is focused while the tab key is pressed, focus should
switch to the first autocomplete item. In Safari / WebKit, this works
without further modification.

In Firefox, we simply lose focus. To mitigate this, we have to manually
override the dropdown's visibility, focus the first item, and remove our
override.